#.
- 데이터 테이블 에러 대응 (안전장치)
- 유닛 이동속도 반영

+++ Assets/02. Scripte/DataLibrary/Main/DataTable/Data_Monster.cs
@@ -33,7 +33,7 @@ namespace IdleGame.Data
         /// <summary>
         /// [데이터] 몬스터의 이동속도입니다.
         /// </summary>
-        public int speed = -1;
+        public float speed = -1;
 
         /// <summary>
         /// [데이터] 몬스터의 최대 체력입니다.
